Add refresh token expiry hook (#15)
This adds a hook which is used to determine whether an access token is still valid, or should be expired.

The default implementation never expires refresh tokens which (mostly) matches behaviour of public OPs.

In the event a refresh token is expired it is immediately removed from the database, along with the access token.

An issued_at field is added to the token model and passed into the hook to make it possible to implement a simple time-based expiry. This field is not used elsewhere in the application.